# docs/swagger.json
{
  "x-generator": "NSwag v14.5.0.0 (NJsonSchema v11.4.0.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "PodMD API",
    "version": "1.0.0"
  },
  "paths": {
    "/api-keys": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "operationId": "GetApiKey",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/clusters": {
      "get": {
        "tags": [
          "Clusters"
        ],
        "operationId": "GetClusters",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClusterResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clusters"
        ],
        "operationId": "RegisterCluster",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClusterRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterResponse"
                }
              }
            }
          }
        }
      }
    },
    "/clusters/{clusterGuid}": {
      "put": {
        "tags": [
          "Clusters"
        ],
        "operationId": "UpdateCluster",
        "parameters": [
          {
            "name": "clusterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClusterRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clusters"
        ],
        "operationId": "UnregisterCluster",
        "parameters": [
          {
            "name": "clusterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/clusters/{clusterGuid}/knowledge-bases": {
      "get": {
        "tags": [
          "Clusters"
        ],
        "operationId": "GetClusterKnowledgeBases",
        "parameters": [
          {
            "name": "clusterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeBaseResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clusters/{clusterGuid}/knowledge-bases/{knowledgeBaseGuid}": {
      "post": {
        "tags": [
          "Clusters"
        ],
        "operationId": "LinkClusterToKnowledgeBase",
        "parameters": [
          {
            "name": "clusterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "knowledgeBaseGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Clusters"
        ],
        "operationId": "UnlinkClusterFromKnowledgeBase",
        "parameters": [
          {
            "name": "clusterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "knowledgeBaseGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/clusters/{clusterGuid}/logs/recommend-fixes": {
      "post": {
        "tags": [
          "Clusters"
        ],
        "operationId": "RecommendFixes",
        "parameters": [
          {
            "name": "clusterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HowToFixRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TroubleshootingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-bases": {
      "get": {
        "tags": [
          "Knowledge Bases"
        ],
        "operationId": "GetKnowledgeBases",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeBaseResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Knowledge Bases"
        ],
        "operationId": "CreateKnowledgeBase",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKnowledgeBaseRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-bases/{knowledgeBaseGuid}": {
      "get": {
        "tags": [
          "Knowledge Bases"
        ],
        "operationId": "GetKnowledgeBase",
        "parameters": [
          {
            "name": "knowledgeBaseGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Knowledge Bases"
        ],
        "operationId": "DeleteKnowledgeBase",
        "parameters": [
          {
            "name": "knowledgeBaseGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/knowledge-bases/{knowledgeBaseGuid}/resources": {
      "get": {
        "tags": [
          "Knowledge Bases"
        ],
        "operationId": "GetKnowledgeBaseResources",
        "parameters": [
          {
            "name": "knowledgeBaseGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResourceResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Knowledge Bases"
        ],
        "operationId": "UploadResource",
        "parameters": [
          {
            "name": "knowledgeBaseGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-bases/{knowledgeBaseGuid}/resources/{resourceGuid}": {
      "post": {
        "tags": [
          "Knowledge Bases"
        ],
        "operationId": "ReplaceResource",
        "parameters": [
          {
            "name": "knowledgeBaseGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "resourceGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Knowledge Bases"
        ],
        "operationId": "DeleteResource",
        "parameters": [
          {
            "name": "knowledgeBaseGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "resourceGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiKeyResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "apiKey": {
            "type": "string"
          },
          "lookupName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ClusterResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "guid": {
            "type": "string",
            "format": "guid"
          },
          "host": {
            "type": "string"
          }
        }
      },
      "CreateClusterRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "host": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "UpdateClusterRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "host": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "KnowledgeBaseResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "guid": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "TroubleshootingResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorGroup"
            }
          }
        }
      },
      "ErrorGroup": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "general_message": {
            "type": "string"
          },
          "occurrences": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "solutions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Solution"
            }
          }
        }
      },
      "Solution": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "steps": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/SolutionStep"
            }
          }
        }
      },
      "SolutionStep": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "explanation": {
            "type": "string",
            "nullable": true
          },
          "command": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "HowToFixRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "namespace": {
            "type": "string"
          },
          "pod": {
            "type": "string"
          },
          "tail": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sinceTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CreateKnowledgeBaseRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ResourceResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "guid": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          }
        }
      }
    }
  }
}